requirements.txt
Flask
flask-cors
langchain
langchain-groq
sentence-transformers
faiss-cpu
python-dotenv
gTTS
streamlit
requests
pyttsx3
SpeechRecognition
